moved setting folder from otherComponent to Component root
