--- App/components/otherComponents/setting/password/PasswordSetting.tsx
@@ -1,11 +1,0 @@
-import { PasswordSettingContainer } from "./password.styles";
-
-const PasswordSetting = () => {
-  return (
-    <PasswordSettingContainer>
-      <h1>PASSWORD SETTING GOES HERE</h1>
-    </PasswordSettingContainer>
-  );
-};
-
-export default PasswordSetting;
