+++ App/components/setting/AccountInfo/AccountInfo.tsx
@@ -7,7 +7,7 @@ import {
   ProfileButton,
   ProfileContainer,
 } from "./AccountInfo.styles";
-import profile from "../../../../assets/images/profile.png";
+import profile from "../../../assets/images/profile.png";
 import Image from "next/image";
 import { Icon } from "@iconify/react";
