+++ App/components/setting/PswdSetting/PswdSetting.styles.ts
@@ -1,10 +1,10 @@
 import styled from "styled-components";
 
-const PasswordSettingContainer = styled.div`
+const PswdSettingContainer = styled.div`
   display: grid;
   place-content: center;
   font-size: 20px;
   color: #fff;
 `;
 
-export { PasswordSettingContainer };
+export { PswdSettingContainer };
