+++ App/components/setting/PswdSetting/PswdSetting.tsx
@@ -1,0 +1,11 @@
+import { PswdSettingContainer } from "./PswdSetting.styles";
+
+const PswdSetting = () => {
+  return (
+    <PswdSettingContainer>
+      <h1>PASSWORD SETTING GOES HERE</h1>
+    </PswdSettingContainer>
+  );
+};
+
+export default PswdSetting;
